Initial Commit
Made a basic prototype for dynamic Spatial UI
Interaction 1: When an area of interest is identified, the user is given a visual affordance (Public )
Interaction 2: When the user walks toward the area of interest, more information is revealed. (Personal)
Interaction 3: When the user reaches the area of interest, an interactive window will be displayed. (Personal)